% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flrpkgGraphics-package.R
\docType{package}
\name{flrpkgGraphics}
\alias{flrpkgGraphics-package}
\alias{flrpkgGraphics}
\title{Graphics utilities used by Ephel}
\description{
Graphics utilities used by Ephel.
}
\details{
Custom functions and objects used for graphics such as themes, plots, etc.
}
\section{flrpkgGraphics functions}{

The functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FrankLef/flrpkgGraphics}
  \item Report bugs at \url{https://github.com/FrankLef/flrpkgGraphics/issues}
}

}
\author{
\strong{Maintainer}: François Lefebvre \email{flefebvre01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
