% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_dens_hrdl.R
\name{ggp_den_hrdl}
\alias{ggp_den_hrdl}
\title{Density plot for hurdle analysis}
\usage{
ggp_den_hrdl(data, x_var, flag_var, vals, fun = identity)
}
\arguments{
\item{data}{Data.fame.}

\item{x_var}{String to identify the x variable.}

\item{flag_var}{String to identify flag variable}

\item{vals}{Named character() with colors.}

\item{fun}{Function used to transform the data. Default is \code{identity}.
Very common to use \code{ceiling(flrpkgTools::expm1s())}.}
}
\value{
Object of class \code{patchwork, gg}.
}
\description{
Density plot for hurdle analysis.
}
\details{
Density plot for hurdle analysis using a custom theme.
}
\examples{
\dontrun{
TODO
}
}
