% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_dens_hrdl_many.R
\name{ggp_den_hrdl_many}
\alias{ggp_den_hrdl_many}
\title{Many density plots for hurdle analysis}
\usage{
ggp_den_hrdl_many(data, cols, vals, titles)
}
\arguments{
\item{data}{Data.fame.}

\item{cols}{Named character vector where the names are the value variables
and the string are the name of the falg variables.}

\item{vals}{Named character() with colors.}

\item{titles}{List of titles used by \code{labs}.}
}
\value{
Object of class \code{gg}.
}
\description{
Many density plots for hurdle analysis.
}
\details{
Many density plots for hurdle analysis all joined together using
\pkg{patchwork}.
}
\examples{
\dontrun{
TODO
}
}
