% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_noncompliant_tree.R
\name{ggp_noncompliant_tree}
\alias{ggp_noncompliant_tree}
\title{Create a tree plot of non-compliant data}
\usage{
ggp_noncompliant_tree(obj, fill_var, fill_colrs)
}
\arguments{
\item{obj}{Object of class \code{constparty}.}

\item{fill_var}{String with name of the fill variable.}

\item{fill_colrs}{Palette of fill colors.}
}
\value{
Object of class \code{gg} created by \code{ggparty::ggparty()}.
}
\description{
Create a tree plot of non-compliant data.
}
\details{
Create a tree plot of conditional tree of non-compliant with colors by
cluster. The conditional tree is computed with \code{partykit::ctree}.
}
\examples{
\dontrun{
TODO
}
}
