% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_noncompliant_tree.R
\name{ggp_noncompliant_tree_title}
\alias{ggp_noncompliant_tree_title}
\title{Create the title used by \code{ggp_noncompliant_tree}}
\usage{
ggp_noncompliant_tree_title(data, prune_var, main, submain)
}
\arguments{
\item{data}{Data.frame that was used to create the conditional tree.}

\item{prune_var}{String with the name of the prune variable.}

\item{main}{String with text for the main title.}

\item{submain}{String with text for the subtitle.}
}
\value{
List with \code{title} and \code{subtitle}.
}
\description{
Create the title used by \code{ggp_noncompliant_tree}.
}
\details{
The subtitle will include the number of observations.
}
\examples{
\dontrun{
TODO
}

}
