% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_pairs.R
\name{ggp_pairs}
\alias{ggp_pairs}
\title{Matrix of plots with a given data set}
\usage{
ggp_pairs(data, columns, color_var, pal, titles)
}
\arguments{
\item{data}{Data.frame.}

\item{columns}{Character() with name of columns to plot.}

\item{color_var}{Variable to use to color by group.}

\item{pal}{Color to use by \code{color_var}.}

\item{titles}{List of titles.}
}
\value{
Object of class \code{gg}.
}
\description{
Matrix of plots with a given data set.
}
\details{
Matrix of plots with density plot on the diagonal, scatter plots on
lower triangle and correlations on upper triangle.
}
\examples{
\dontrun{
TODO
}
}
