% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_scatter_addval.R
\name{ggp_scatter_addval}
\alias{ggp_scatter_addval}
\title{Scatter plot for added value analysis}
\usage{
ggp_scatter_addval(data, x_var, y_var, group_var, pal, fun, titles)
}
\arguments{
\item{data}{Data.frame.}

\item{x_var}{String with name of the x axis.}

\item{y_var}{String with name of the y axis.}

\item{group_var}{String with name of the group variable.}

\item{pal}{Color palette for the group.}

\item{fun}{Function to convert axes' labels.}

\item{titles}{List of titles.}
}
\value{
Object of class \code{gg}.
}
\description{
Scatter plot for added value analysis.
}
\details{
Scatter plot for added value analysis with a custom format. Labels on the
axes are converted to the natural scale.
}
\examples{
\dontrun{
TODO
}
}
