% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_thm_clean.R
\name{ggp_thm_clean}
\alias{ggp_thm_clean}
\title{\pkg{ggplot2} \emph{clean theme} with light nice colors}
\usage{
ggp_thm_clean(
  base_theme,
  base_family = "Noto Sans Display",
  legend = list(position = "none"),
  title_colr = "midnightblue",
  strip_colr = list(bgfill = "gainsboro"),
  strip_text = list(face = "bold", color = "darkblue"),
  panel_colr = list(fill = "snow", color = "snow"),
  plot_colr = list(fill = "ghostwhite", color = "ghostwhite")
)
}
\arguments{
\item{base_theme}{Function returning the base theme build upon.}

\item{base_family}{String of font family used.}

\item{legend}{List with legend parameters.}

\item{title_colr}{String with title's color.}

\item{strip_colr}{List of parameters for facets' strips colors.}

\item{strip_text}{List of parameters for facets' strips text.}

\item{panel_colr}{List of colors for the panel.}

\item{plot_colr}{List of colors for the plot.}
}
\value{
Object of class \code{theme}.
}
\description{
\pkg{ggplot2} \emph{clean theme} with light nice colors.
}
\details{
The clean format uses light colors and reduces visual distraction much like
the \emph{tufte} ethics of graphics.
}
\examples{
\dontrun{
TODO
}
}
