% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggp_thm_hrclean.R
\name{ggp_thm_hrclean}
\alias{ggp_thm_hrclean}
\title{GGplot \emph{clean theme} with \code{hrbrthemes::theme_ipsum}}
\usage{
ggp_thm_hrclean(
  base_family = "Noto Sans Display",
  legend = list(position = "none"),
  title_colr = "midnightblue",
  strip_colr = list(bgfill = "gainsboro"),
  strip_text = list(face = "bold", color = "darkblue"),
  panel_colr = list(fill = "snow", color = "snow"),
  plot_colr = list(fill = "ghostwhite", color = "ghostwhite")
)
}
\arguments{
\item{base_family}{String of font family used.}

\item{legend}{List with legend parameters.}

\item{title_colr}{String with title's color.}

\item{strip_colr}{List of parameters for facets' strips colors.}

\item{strip_text}{List of parameters for facets' strips text.}

\item{panel_colr}{List of colors for the panel.}

\item{plot_colr}{List of colors for the plot.}
}
\value{
Object of class \code{theme}.
}
\description{
GGplot \emph{clean theme} with \code{hrbrthemes::theme_ipsum}.
}
\details{
Use the \code{hrbrthemes::theme_ipsum} with a clean format.
}
\examples{
\dontrun{
TODO
}
}
