% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paltr_colrs_scales.R
\name{paltr_colrs_scales}
\alias{paltr_colrs_scales}
\title{Create a named vector of colors}
\usage{
paltr_colrs_scales(pal, vals, direction = 1L)
}
\arguments{
\item{pal}{Name of palette available in \pkg{paletteer}.}

\item{vals}{Character vector of values.}

\item{direction}{Direction of palette. See \pkg{paletteer} for details.}
}
\value{
Named vector with color names.
}
\description{
Create a named vector of colors.
}
\details{
The character vector is in the form \code{c("value" = "color")}. The colors
come from a palette provided by \pkg{paletter}.
}
\examples{
\dontrun{
TODO
}
}
