% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply_scatter3D_valid.R
\name{ply_scatter3D_valid_axis}
\alias{ply_scatter3D_valid_axis}
\title{Create the axis used by \code{ply_scatter3D_valid}}
\usage{
ply_scatter3D_valid_axis(data, var, title, colrs, fun = identity)
}
\arguments{
\item{data}{Data.frame.}

\item{var}{String with name of the variable used for the axis.}

\item{title}{String with the axis' label.}

\item{colrs}{List with colors.}

\item{fun}{Function used to tranform values on the scale. default is i
\code{dentity()}.}
}
\value{
List of arguments used by \code{plotly::layout}.
}
\description{
Create the axis used by \code{ply_scatter3D_valid}.
}
\details{
Create the axis used by \code{ply_scatter3D_valid} with labels formatted
using \code{fun}.
}
\examples{
\dontrun{
TODO
}

}
\seealso{
\link{ply_scatter3D_valid}
}
