% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply_scatter3D_valid.R
\name{ply_scatter3D_valid_title}
\alias{ply_scatter3D_valid_title}
\title{Create the string of title ans subtitle for \code{ply_scatter3D_valid}}
\usage{
ply_scatter3D_valid_title(data, main, submain)
}
\arguments{
\item{data}{Data.frame.}

\item{main}{String used for the title.}

\item{submain}{String used for the subtitle.}
}
\value{
String with html tags..
}
\description{
Create the string of title ans subtitle for \code{ply_scatter3D_valid}.
}
\details{
Create the string of title ans subtitle for \code{ply_scatter3D_valid} by
combining \code{main} and \code{submain} with line breaks in a html string.
}
\examples{
\dontrun{
TODO
}
}
\seealso{
\link{ply_scatter3D_valid}
}
