% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_titles.R
\name{write_titles}
\alias{write_titles}
\title{Create a list of titles and axis labels}
\usage{
write_titles(data, title, subtitle = NULL, is_main = FALSE, ...)
}
\arguments{
\item{data}{Data frame.}

\item{title}{String with the title.}

\item{subtitle}{String with the subtitle.}

\item{is_main}{TRUE: The title and subtitle will be put together with a line
break as one main title. Useful when using \pkg{plotly} for example. FALSE is
is the default value.}

\item{...}{Paired variables.}
}
\value{
List of titles and axis labels.
}
\description{
Create a list of titles and axis labels.
}
\details{
The list will only include the axis for which a variable is goven in
\code{...}. The labels are managed with the package \pkg{labelled}.
}
\examples{
\dontrun{
TODO
}
}
