% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply_scatter3D_valid.R
\name{ply_scatter3D_valid}
\alias{ply_scatter3D_valid}
\title{Create a 3D plot with \pkg{plotly} to show valid vs invalid data}
\source{
\url{https://plotly.com/r/3d-axes/}.
}
\usage{
ply_scatter3D_valid(
  data,
  x_var,
  y_var,
  z_var,
  color_var,
  colrs,
  fun = identity,
  titles
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_var}{String with name of x variable.}

\item{y_var}{String with name of y variable.}

\item{z_var}{String with name of z variable.}

\item{color_var}{String with name of color variable.}

\item{colrs}{List with colors.}

\item{fun}{Function used to tranform values on the scale. default is i
\code{dentity()}.}

\item{titles}{List of titles.}
}
\value{
Object of class \code{"plotly", "htmlwidget"}.
}
\description{
Create a 3D plot with \pkg{plotly} to show valid vs invalid data.
}
\details{
The 3D graph will be colored based on the \code{color_var} to differentiate
the valid vs invalid data.
}
\examples{
\dontrun{
TODO
}
}
